requirements.txt
langchain
python-dotenv
langchain-openai
ipykernel
beautifulsoup4 # this library is used for web scraping
langchain_community
faiss-cpu # this is used for vector storage
streamlit
langchain_groq
langchain_core